feat: final ordering done
