+++ apps/client/src/components/core/index.tsx
@@ -5,6 +5,6 @@ export * from "./Navbar";
 export * from "./Image";
 import ProductCard from "./cards/ProductCard";
 export * from "./StatusViewer";
-export * from "./PaymentModalButton";
+export * from "./Payment/PaymentModalButton";
 
 export { ProductCard };
