+++ apps/server/src/modules/order/models/index.ts
@@ -3,3 +3,4 @@ export * from './Order';
 export * from './Track';
 export * from './TackLog';
 export * from './UserAddress';
+export * from './PaymentMethod';
