+++ apps/server/src/modules/order/order.module.ts
@@ -6,10 +6,11 @@ import { ShopModule } from '../shop/shop.module';
 import { OrderResolver } from './resolver/order.resolver';
 import { AddressService } from './services/Address.service';
 import { AddressResolver } from './resolver/address.resolver';
+import { PaymentResolver } from './resolver/payment.resolver';
 
 @Module({
   imports: [ShopModule, TypeOrmModule.forFeature(Object.values(models))],
-  providers: [OrderService, OrderResolver, AddressService, AddressResolver],
+  providers: [OrderService, OrderResolver, PaymentResolver, AddressService, AddressResolver],
   exports: [OrderService],
 })
 export class OrderModule {}
